Create SpamBuilder.Java
Parses through a text file, and chunks it, These chunks are then used to create "say" binds. The class also creates a bind loop chain on whatever key has been chosen in the PinkuBindSpam class. 

This means that whatever key is chosen will rebind itself to say the next chunk of the text in the chat the next time it is pressed, and bind itself to the first chunk once the loop has concluded.